fix(encryption): Prevent unencrypted empty array in tags field
When updating a transaction, if the incoming data from Plaid did not
contain a category field, the tags field in the database was being
set to an unencrypted empty array. This caused decryption errors when
the application later tried to read this field.

This commit fixes the issue by explicitly setting the tags field to
null if the category field is not present in the incoming data.

Also adds a debug script to help identify corrupted encrypted data.